move definition of prepareCreateDomainInput function
