+++ services/elasticsearch/create_worker_test.go
@@ -20,6 +20,7 @@ import (
 	"github.com/cloud-gov/aws-broker/helpers"
 	"github.com/cloud-gov/aws-broker/helpers/request"
 	"github.com/cloud-gov/aws-broker/testutil"
+	"github.com/go-test/deep"
 	"github.com/google/uuid"
 	"github.com/riverqueue/river"
 )
@@ -494,6 +495,203 @@ func TestCreateWorkerWork(t *testing.T) {
 			}
 			if message.JobState.State != test.expectedState {
 				t.Fatalf("expected %s, got %s", test.expectedState, message.JobState.State)
+			}
+		})
+	}
+}
+
+func TestPrepareCreateDomainInput(t *testing.T) {
+	testCases := map[string]struct {
+		esInstance     *ElasticsearchInstance
+		accessPolicy   string
+		expectedParams *opensearch.CreateDomainInput
+	}{
+		"data count of 1": {
+			esInstance: &ElasticsearchInstance{
+				Domain:                     "test-domain",
+				DataCount:                  1,
+				SubnetID2AZ2:               "az-2",
+				SecGroup:                   "group-1",
+				EncryptAtRest:              false,
+				VolumeSize:                 10,
+				VolumeType:                 "gp3",
+				InstanceType:               "m5.2xlarge.search",
+				NodeToNodeEncryption:       true,
+				AutomatedSnapshotStartHour: 0,
+				Tags: map[string]string{
+					"foo": "bar",
+				},
+			},
+			accessPolicy: "fake-access-policy",
+			expectedParams: &opensearch.CreateDomainInput{
+				DomainName:     aws.String("test-domain"),
+				AccessPolicies: aws.String("fake-access-policy"),
+				VPCOptions: &opensearchTypes.VPCOptions{
+					SubnetIds:        []string{"az-2"},
+					SecurityGroupIds: []string{"group-1"},
+				},
+				DomainEndpointOptions: &opensearchTypes.DomainEndpointOptions{
+					EnforceHTTPS: aws.Bool(true),
+				},
+				EBSOptions: &opensearchTypes.EBSOptions{
+					EBSEnabled: aws.Bool(true),
+					VolumeSize: aws.Int32(int32(10)),
+					VolumeType: opensearchTypes.VolumeTypeGp3,
+				},
+				ClusterConfig: &opensearchTypes.ClusterConfig{
+					InstanceType:  opensearchTypes.OpenSearchPartitionInstanceTypeM52xlargeSearch,
+					InstanceCount: aws.Int32(int32(1)),
+				},
+				SnapshotOptions: &opensearchTypes.SnapshotOptions{
+					AutomatedSnapshotStartHour: aws.Int32(int32(0)),
+				},
+				NodeToNodeEncryptionOptions: &opensearchTypes.NodeToNodeEncryptionOptions{
+					Enabled: aws.Bool(true),
+				},
+				EncryptionAtRestOptions: &opensearchTypes.EncryptionAtRestOptions{
+					Enabled: aws.Bool(false),
+				},
+				TagList: []opensearchTypes.Tag{
+					{
+						Key:   aws.String("foo"),
+						Value: aws.String("bar"),
+					},
+				},
+			},
+		},
+		"data count is greater than 1": {
+			esInstance: &ElasticsearchInstance{
+				Domain:                     "test-domain",
+				DataCount:                  2,
+				SubnetID3AZ1:               "az-3",
+				SubnetID4AZ2:               "az-4",
+				SecGroup:                   "group-1",
+				EncryptAtRest:              false,
+				VolumeSize:                 10,
+				VolumeType:                 "gp3",
+				InstanceType:               "m5.2xlarge.search",
+				NodeToNodeEncryption:       true,
+				AutomatedSnapshotStartHour: 0,
+			},
+			accessPolicy: "fake-access-policy",
+			expectedParams: &opensearch.CreateDomainInput{
+				DomainName:     aws.String("test-domain"),
+				AccessPolicies: aws.String("fake-access-policy"),
+				VPCOptions: &opensearchTypes.VPCOptions{
+					SubnetIds:        []string{"az-3", "az-4"},
+					SecurityGroupIds: []string{"group-1"},
+				},
+				DomainEndpointOptions: &opensearchTypes.DomainEndpointOptions{
+					EnforceHTTPS: aws.Bool(true),
+				},
+				EBSOptions: &opensearchTypes.EBSOptions{
+					EBSEnabled: aws.Bool(true),
+					VolumeSize: aws.Int32(int32(10)),
+					VolumeType: opensearchTypes.VolumeTypeGp3,
+				},
+				ClusterConfig: &opensearchTypes.ClusterConfig{
+					InstanceType:         opensearchTypes.OpenSearchPartitionInstanceTypeM52xlargeSearch,
+					InstanceCount:        aws.Int32(int32(2)),
+					ZoneAwarenessEnabled: aws.Bool(true),
+					ZoneAwarenessConfig: &opensearchTypes.ZoneAwarenessConfig{
+						AvailabilityZoneCount: aws.Int32(int32(2)),
+					},
+				},
+				SnapshotOptions: &opensearchTypes.SnapshotOptions{
+					AutomatedSnapshotStartHour: aws.Int32(int32(0)),
+				},
+				NodeToNodeEncryptionOptions: &opensearchTypes.NodeToNodeEncryptionOptions{
+					Enabled: aws.Bool(true),
+				},
+				EncryptionAtRestOptions: &opensearchTypes.EncryptionAtRestOptions{
+					Enabled: aws.Bool(false),
+				},
+			},
+		},
+		"audit + error logs enable FGAC and log publishing": {
+			esInstance: &ElasticsearchInstance{
+				Domain:                     "test-domain",
+				DataCount:                  2,
+				SubnetID3AZ1:               "az-3",
+				SubnetID4AZ2:               "az-4",
+				SecGroup:                   "group-1",
+				EncryptAtRest:              true,
+				VolumeSize:                 10,
+				VolumeType:                 "gp3",
+				InstanceType:               "m5.2xlarge.search",
+				NodeToNodeEncryption:       true,
+				AutomatedSnapshotStartHour: 0,
+				ErrorLogsEnabled:           true,
+				AuditLogsEnabled:           true,
+				AdvancedSecurityEnabled:    true,
+				IamUserARN:                 "arn:aws-us-gov:iam::123456789012:user/test-domain",
+				ErrorLogsGroupARN:          "arn:aws-us-gov:logs:us-gov-west-1:123456789012:log-group:/aws/OpenSearchService/domains/test-domain/application-logs",
+				AuditLogsGroupARN:          "arn:aws-us-gov:logs:us-gov-west-1:123456789012:log-group:/aws/OpenSearchService/domains/test-domain/audit-logs",
+			},
+			accessPolicy: "fake-access-policy",
+			expectedParams: &opensearch.CreateDomainInput{
+				DomainName:     aws.String("test-domain"),
+				AccessPolicies: aws.String("fake-access-policy"),
+				VPCOptions: &opensearchTypes.VPCOptions{
+					SubnetIds:        []string{"az-3", "az-4"},
+					SecurityGroupIds: []string{"group-1"},
+				},
+				DomainEndpointOptions: &opensearchTypes.DomainEndpointOptions{
+					EnforceHTTPS: aws.Bool(true),
+				},
+				EBSOptions: &opensearchTypes.EBSOptions{
+					EBSEnabled: aws.Bool(true),
+					VolumeSize: aws.Int32(int32(10)),
+					VolumeType: opensearchTypes.VolumeTypeGp3,
+				},
+				ClusterConfig: &opensearchTypes.ClusterConfig{
+					InstanceType:         opensearchTypes.OpenSearchPartitionInstanceTypeM52xlargeSearch,
+					InstanceCount:        aws.Int32(int32(2)),
+					ZoneAwarenessEnabled: aws.Bool(true),
+					ZoneAwarenessConfig: &opensearchTypes.ZoneAwarenessConfig{
+						AvailabilityZoneCount: aws.Int32(int32(2)),
+					},
+				},
+				SnapshotOptions: &opensearchTypes.SnapshotOptions{
+					AutomatedSnapshotStartHour: aws.Int32(int32(0)),
+				},
+				NodeToNodeEncryptionOptions: &opensearchTypes.NodeToNodeEncryptionOptions{
+					Enabled: aws.Bool(true),
+				},
+				EncryptionAtRestOptions: &opensearchTypes.EncryptionAtRestOptions{
+					Enabled: aws.Bool(true),
+				},
+				AdvancedSecurityOptions: &opensearchTypes.AdvancedSecurityOptionsInput{
+					Enabled:                     aws.Bool(true),
+					InternalUserDatabaseEnabled: aws.Bool(false),
+					MasterUserOptions: &opensearchTypes.MasterUserOptions{
+						MasterUserARN: aws.String("arn:aws-us-gov:iam::123456789012:user/test-domain"),
+					},
+				},
+				LogPublishingOptions: map[string]opensearchTypes.LogPublishingOption{
+					"AUDIT_LOGS": {
+						CloudWatchLogsLogGroupArn: aws.String("arn:aws-us-gov:logs:us-gov-west-1:123456789012:log-group:/aws/OpenSearchService/domains/test-domain/audit-logs"),
+						Enabled:                   aws.Bool(true),
+					},
+					"ES_APPLICATION_LOGS": {
+						CloudWatchLogsLogGroupArn: aws.String("arn:aws-us-gov:logs:us-gov-west-1:123456789012:log-group:/aws/OpenSearchService/domains/test-domain/application-logs"),
+						Enabled:                   aws.Bool(true),
+					},
+				},
+			},
+		},
+	}
+	for name, test := range testCases {
+		t.Run(name, func(t *testing.T) {
+			params, err := prepareCreateDomainInput(
+				test.esInstance,
+				test.accessPolicy,
+			)
+			if err != nil {
+				t.Fatal(err)
+			}
+			if diff := deep.Equal(params, test.expectedParams); diff != nil {
+				t.Error(diff)
 			}
 		})
 	}
